Resolve warnings regarding missing casts
- As NSInteger and NSUInteger lengths vary based upon 32/64 bit architectures,
Xcode requires these to be explicitly casted when using in format specifiers.

+++ QredoSDK/Internal/Rendezvous/QredoAuthenticatedRendezvousTag.m
@@ -65,8 +65,8 @@
     NSArray *splitTagParts = [fullTag componentsSeparatedByString:@"@"];
     NSUInteger separatorCount = splitTagParts.count - 1;
     if (separatorCount != 1) {
-        NSString *message = [NSString stringWithFormat:@"Invalid number (%ld) of @ characters present. Require exactly 1. Full tag: '%@'",
-                             separatorCount, fullTag];
+        NSString *message = [NSString stringWithFormat:@"Invalid number (%lu) of @ characters present. Require exactly 1. Full tag: '%@'",
+                             (unsigned long)separatorCount, fullTag];
         LogError(@"%@", message);
         if (error) {
             *error = createError(QredoAuthenticatedRendezvousTagErrorMalformedTag, message);
